+++ kernel/src/syscall.rs
@@ -86,8 +86,8 @@ pub enum Syscall {
     /// is the subscribe identifier, `callback_ptr` is callback pointer,
     /// and `appdata` is the application data.
     Subscribe {
-        driver_number: usize,
-        subdriver_number: usize,
+        driver_number: u32,
+        subdriver_number: u32,
         callback_ptr: *mut (),
         appdata: usize,
     },
@@ -160,8 +160,8 @@ impl Syscall {
                 address: r1 as *mut u8,
             }),
             Ok(SyscallClass::Subscribe) => Some(Syscall::Subscribe {
-                driver_number: r0,
-                subdriver_number: r1,
+                driver_number: r0 as u32,
+                subdriver_number: r1 as u32,
                 callback_ptr: r2 as *mut (),
                 appdata: r3,
             }),
